generator_requirements.txt
requests==2.31.0
python-dotenv==1.0.0
replicate==0.15.4
